Add posts data file to organize different data types
